chore: clean scheme setup
On-behalf-of: SAP <matthias.bruns@sap.com>
